requirements.txt
absl-py==0.7.0
astor==0.7.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
gast==0.2.2
gevent==1.4.0
google-cloud-storage==1.31.0
greenlet==0.4.15
grpcio==1.19.0
gunicorn==20.0.4
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Keras-Applications==1.0.8
Keras-Preprocessing==1.0.9
Markdown==3.0.1
MarkupSafe==1.1.1
mock==2.0.0
numpy==1.16.2
opensimplex==0.3
pbr==5.1.2
Pillow==5.4.1
protobuf==3.6.1
requests==2.24.0
six==1.12.0
tensorflow-gpu==1.15.0
termcolor==1.1.0
tqdm==4.48.0
urllib3[secure]==1.25.7
Werkzeug==0.14.1
wget==3.2
runway-python
